third_party/blink/web_tests/wpt_internal/prerender/session-history.https.html

<!DOCTYPE html>
<title>Test history.length</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
    async function runTestInPrerender(testName) {
      const prerender = "session-history-prerender.https.html";
      const testChannel = new BroadcastChannel(`test-channel-${testName}`);
      const result = new Promise((resolve) => {
        testChannel.addEventListener(
          "message",
          (e) => {
            testChannel.close();
            resolve(e.data);
          },
          { once: true }
        );
      });

      // Run test in a new window for test isolation.
      window.open(
        `./resources/session-history-initiator.https.html?prerender=${prerender}&testName=${testName}`,
        "_blank",
        "noopener"
      );
      return result;
    }

    // This will activate the prerendered context created in runTestInPrerender
    // and then run the post-activation variation of |testName|.
    async function runTestInActivatedPage(testName) {
      const testChannel = new BroadcastChannel(`test-channel-${testName}`);

      const result = new Promise((resolve) => {
        testChannel.addEventListener(
          "message",
          (e) => {
            testChannel.close();
            resolve(e.data);
          },
          { once: true }
        );
      });

      testChannel.postMessage("activate");
      return result;
    }

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryPushStateInPrerender"),
        "Passed"
      );
    }, "history.pushState navigates independently with replacement in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryReplaceStateInPrerender"),
        "Passed"
      );
    }, "history.replaceState navigates independently in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testLocationAssignInPrerender"),
        "Passed"
      );
    }, "location.assign navigates independently with replacement in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testLocationReplaceInPrerender"),
        "Passed"
      );
    }, "location.replace navigates independently in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testSetLocationHrefInPrerender"),
        "Passed"
      );
    }, "Setting location.href navigates independently with replacement in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testSyntheticAnchorClickInPrerender"),
        "Passed"
      );
    }, "Synthetic anchor click navigates independently with replacement in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryGoInPrerender"),
        "Passed"
      );
    }, "history.go(0) navigates independently in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testLocationReloadInPrerender"),
        "Passed"
      );
    }, "location.reload() navigates independently in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryLengthInPrerender"),
        "Passed"
      );
      assert_equals(
        await runTestInActivatedPage("testHistoryLengthInPrerender"),
        "Passed"
      );
    }, "history.length should be updated after activation");

  </script>
</body>
